% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_filter.R
\name{distance_filter}
\alias{distance_filter}
\title{Distance threshold filter}
\usage{
distance_filter(x, d = 5, select = NULL)
}
\arguments{
\item{x}{'track' object.}

\item{d}{Distance threshold, same units original data.}

\item{select}{character. The filter is applied only in these dives.}
}
\description{
Only keep points after a distance \code{d} has been reached.
Check \href{https://www.gpsvisualizer.com/tutorials/track_filters.html}{here} for a more detailed
explanation of how the filter works.
}
