% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_filter.R
\name{fill_time_gaps}
\alias{fill_time_gaps}
\title{Create a point every second}
\usage{
fill_time_gaps(x, type = "locf")
}
\arguments{
\item{x}{'track' object.}

\item{type}{interpolation to be used. One of 'locf', 'approx' or 'spline'.}
}
\description{
Fill gaps in the timestamp of track, generating a point every second.
Missing points are interpolated based on \code{type}.
}
\note{
This should be the last filter used in the track data.
}
