% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_filter.R
\name{filter_depth}
\alias{filter_depth}
\title{Filter points based on depth}
\usage{
filter_depth(x, depth = -5, select = NULL)
}
\arguments{
\item{x}{'track' object.}

\item{depth}{threshold to filter points, same units original data.
Points with depth above this value will be removed.}

\item{select}{character. The filter is applied only in these dives.}
}
\description{
Remove points where depth is greater than \code{depth}.
}
