% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_filter.R
\name{interpolate_points}
\alias{interpolate_points}
\title{Interpolate points along track}
\usage{
interpolate_points(x, density = 2)
}
\arguments{
\item{x}{'track' object.}

\item{density}{Number of points per distance unit. Passed to \code{\link[sf]{st_line_sample}}.}
}
\description{
Create points along track and assign original timestamp to the new points.
}
