% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_filter.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(x, n = 7, d = 5, select = NULL)
}
\arguments{
\item{x}{'track' object.}

\item{n}{window size to search for outlier.}

\item{d}{threshold to remove outliers, same units original data.}

\item{select}{character. The filter is applied only in these dives.}
}
\description{
Remove points that are far away from adjacent points.
}
\details{
This filter will divide the window in three parts. A median is calculated from both ends,
and if the center value is more distant than \code{d} from the mean, this value is considered an outlier.
This process is repeated for each coordinate.
}
