% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_filter.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Rolling mean filter}
\usage{
rolling_mean(x, n = 15, select = NULL)
}
\arguments{
\item{x}{'track' object.}

\item{n}{window size to calculate the mean.}

\item{select}{character. The filter is applied only in these dives.}
}
\description{
Apply a simple rolling mean filter.
}
