% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_tracks}
\alias{save_tracks}
\title{Save plots of all dives}
\usage{
save_tracks(x, ...)
}
\arguments{
\item{x}{'track' object.}

\item{...}{further arguments passed to \code{plot.track}.}
}
\description{
Plot and save tracks of each dive in a track object as a html file.
}
\details{
Only the \code{plotly} engine is supported.
}
