% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_dives}
\alias{set_dives}
\title{Generate Dive column}
\usage{
set_dives(x, dive_ranges)
}
\arguments{
\item{x}{'track' object.}

\item{dive_ranges}{A data.frame or data.table with columns:
\itemize{
  \item \code{Dive} - Name of the dive.
  \item \code{start} - timestamp of the start of the dive, same format as the \code{Time}
   used to create the track.
  \item \code{end} - timestamp of the end of the dive, same format as \code{start}.
  }}
}
\description{
Associate dives based on start and end timestamps.
}
\details{
This function will overwrite the dives column based on timestamps of start and end.
Any point not contained in the interval of the new dives will
be removed. This function is useful when working with tracks that are not separated by
dives originally, of if you wish to crop a dive in multiple transects.
}
