% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\docType{data}
\name{simcoords}
\alias{simcoords}
\title{Sample of ROV track data}
\format{
A data frame with 4002 rows and 5 variables:
\describe{
  \item{Time}{timestamp of the track, numeric or POSIXct}
  \item{Lon}{Longitude}
  \item{Lat}{Latitude}
  \item{Depth}{Depth}
  \item{Dive}{optional. Character or factor, indicating from which dive is the coordinates}
}
}
\usage{
simcoords
}
\description{
A data.frame containing a sample of coordinates from two ROV dives.
}
\keyword{datasets}
