% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_filter.R
\name{simplify_filter}
\alias{simplify_filter}
\title{Simplify points}
\usage{
simplify_filter(x, dTolerance = 5, preserveTopology = TRUE)
}
\arguments{
\item{x}{'track' object.}

\item{dTolerance, preserveTopology}{Passed to \code{\link[sf]{st_simplify}}.}
}
\description{
Remove points using Douglas-Peuker Algorithm.
}
