% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_filter.R
\name{trim_track}
\alias{trim_track}
\title{Trim start and end of each track}
\usage{
trim_track(x, d = 10, nend = 100, select = NULL)
}
\arguments{
\item{x}{'track' object.}

\item{d}{threshold to remove outliers, same units original data.}

\item{nend}{maximum number of points at each end that can removed.}

\item{select}{character. The filter is applied only in these dives.}
}
\description{
Remove points that are far away in the start or end of a track.
}
\details{
This filter will analyze only the ends of each track. If distance is greater than \code{d},
any previous/following points will be removed.
}
